feat: allow per-chat model selection
